requirements.txt
requests >=2.32.3,<3.0.0
pytest~=8.3.3
jsonschema >=4.23.0,<5.0.0
allure-pytest
allure-python-commons~=2.13.5